% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_giamatti.R
\name{more_giamatti}
\alias{more_giamatti}
\title{Opens a Google Image Search for pictures of Paul Giamatti}
\usage{
more_giamatti()
}
\description{
Opens a Google Image Search for pictures of Paul Giamatti
}
